17404 - Estrutura base do relatório (Grupos e Componentes curriculares)

+++ src/SME.SR.Infra/Dtos/Alunos/AlunoSituacaoDto.cs
@@ -6,7 +6,7 @@ namespace SME.SR.Infra
 {
     public class AlunoSituacaoDto
     {
-        public int CodigoAluno { get; set; }
+        public long CodigoAluno { get; set; }
         public string NomeAluno { get; set; }
         public SituacaoMatriculaAluno CodigoSituacaoMatricula { get; set; }
         public string SituacaoMatricula { get; set; }
